+++ src/java/nghiadh/utils/FileHelpers.java
@@ -12,46 +12,64 @@ import java.io.FileNotFoundException;
 import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.InputStream;
+import java.util.Properties;
 
 /**
  *
  * @author haseo
  */
 public final class FileHelpers {
-    public static void writeImgToServerFile(final String imgFileName,final InputStream fileContent) throws FileNotFoundException, IOException{
-        File imgDir = new File("C:\\PostsImg");
-        if(!imgDir.exists()){
+    private static String IMAGE_SAVING_FOLDER="";
+
+    public static void setIMAGE_SAVING_FOLDER(String IMAGE_SAVING_FOLDER) {
+        FileHelpers.IMAGE_SAVING_FOLDER = IMAGE_SAVING_FOLDER;
+    }
+
+    public static void writeImgToServerFile(final String imgFileName, final InputStream fileContent) throws FileNotFoundException, IOException {
+        File imgDir = new File(IMAGE_SAVING_FOLDER);
+        if (!imgDir.exists()) {
             imgDir.mkdir();
         }
         File imgFile = new File(imgDir, imgFileName);
         FileOutputStream fos = null;
         BufferedInputStream bis = null;
-        try{            
+        try {
             fos = new FileOutputStream(imgFile);
             bis = new BufferedInputStream(fileContent);
-            while(bis.available()>0){
+            while (bis.available() > 0) {
                 fos.write(bis.read());
             }
-        }finally{
-            if(bis!=null)bis.close();
-            if(fos!=null)fos.close();
+        } finally {
+            if (bis != null) {
+                bis.close();
+            }
+            if (fos != null) {
+                fos.close();
+            }
         }
     }
-    public static void copyImgToContextFolder(final String realPath,final String fileName) throws FileNotFoundException, IOException{
-        File inputFile = new File("C:\\PostsImg\\"+fileName);
-        if(inputFile.exists()){
-            File outputFile = new File(realPath+"\\"+fileName);
-            FileInputStream fis = null;
-            FileOutputStream fos = null;
-            try{
-                fis=new FileInputStream(inputFile);
-                fos= new FileOutputStream(outputFile);
-                while(fis.available()>0){
-                    fos.write(fis.read());
+
+    public static void copyImgToContextFolder(final String realPath, final String fileName) throws FileNotFoundException, IOException {
+        FileInputStream fis = null;
+        FileOutputStream fos = null;
+        File inputFile = new File(IMAGE_SAVING_FOLDER+ File.separator + fileName);
+        if (inputFile.exists()) {
+            File outputFile = new File(realPath + File.separator + fileName);
+            if (!outputFile.exists()) {
+                try {
+                    fis = new FileInputStream(inputFile);
+                    fos = new FileOutputStream(outputFile);
+                    while (fis.available() > 0) {
+                        fos.write(fis.read());
+                    }
+                } finally {
+                    if (fos != null) {
+                        fos.close();
+                    }
+                    if (fis != null) {
+                        fis.close();
+                    }
                 }
-            }finally{
-                if(fos!=null)fos.close();
-                if(fis!=null)fis.close();
             }
         }
     }
